fix: harden library, providers, playback, visualizer; bump crates
Correctness
- app: tag every reload/scan result with provider ID + reload generation;
  stale LibraryLoaded/LibraryBatch/LibraryLoadComplete/LibraryScanComplete
  can no longer mutate state after provider switch or superseding reload.
- library/db: enable and verify PRAGMA foreign_keys on every connection,
  wrap migrations in transactions with a version marker, repair legacy
  orphaned playlist_tracks; missing-track cleanup runs in an RAII
  transaction so a failed batch delete rolls back cleanly.
- library/scanner: reconcile missing tracks only under roots that were
  fully and successfully traversed; missing/unreadable/non-directory or
  partially failed roots preserve existing records.
- provider/mpd: create_playlist now does 'save NAME' then 'playlistclear
  NAME' (with 'rm NAME' cleanup on failure), leaving the live queue
  untouched.
- player/http_range_reader: reject unsuccessful initial GETs; nonzero
  seeks require 206 Partial Content with a matching Content-Range;
  short forward-discard preserves the actual consumed position and
  returns UnexpectedEof on early EOF.
- player/engine/decoder: iterative DSD packet loop; bounded reset
  handling, no more unbounded recursion on foreign-track packets.
- autoeq/parser: no more panicking unwraps on network/cache input;
  overflow, non-finite, and out-of-range values return InvalidFormat.
- views/now_playing/visualizer + viz_shader: double-buffered Arc<Vec<u8>>
  frame storage; renderer reuses buffers when uniquely owned, shader
  Arc::clones instead of copying 2 MB per frame; ~126 MB/s of per-frame
  copying eliminated at 30 fps.
- player/engine/conversion: f32_to_i32 scales via f64 so -1.0 maps
  exactly to -i32::MAX (was off by one).

Quality
- lib: drop crate-wide #![allow(dead_code, unused_imports)]; remove all
  newly surfaced unused imports, dead fields, and unused stored AutoEQ
  timeout; add explicit _f32 suffixes on slider steps for the incoming
  float-literal-fallback lint.

Dependencies
- Full cargo update: 45 packages advanced to the latest 1.97-compatible
  resolution (regex 1.13, zbus 5.17, rust-embed 8.12, open 5.4, ...).
- Direct majors bumped: rubato 3.0 -> 4.0, audioadapter-buffers 3.0 -> 4.0;
  resampler migrated to Rubato 4's optional f_cutoff API with the same
  calculated cutoff, quality tiers, channel layout, and fixed-input
  behavior.
- matchit 0.8.4 stays pinned by axum 0.8.9; generic-array 0.14.7 stays
  pinned by crypto-common 0.1.7 via opensubsonic/md-5/secret-service —
  both exact upstream requirements, not local pins.

Tests
- cargo test --workspace --all-targets --all-features: 88 passed, 0 failed.
- Added focused regression coverage for reload generations, DB
  cascades/migrations/orphan repair/multi-chunk rollback, scanner root
  safety, MPD save-then-clear ordering, HTTP range semantics, DSD
  iterative decoding, AutoEQ malformed input, and visualizer
  Arc-shared frame ownership.

+++ src/player/engine/conversion.rs
@@ -34,7 +34,7 @@ pub fn f32_to_i16(val: f32) -> i16 {
 /// Clamp and scale a single f32 sample to `i32` range.
 #[inline]
 pub fn f32_to_i32(val: f32) -> i32 {
-    (val.clamp(-1.0, 1.0) * i32::MAX as f32) as i32
+    (f64::from(val.clamp(-1.0, 1.0)) * f64::from(i32::MAX)) as i32
 }
 
 /// A bounded sample buffer fed from a `SyncSender`/`Receiver` channel.
